+++ map/src/main/java/com/getwemap/example/map/InitialActivity.kt
@@ -1,7 +1,11 @@
 package com.getwemap.example.map
 
 import android.os.Bundle
+import android.view.Menu
+import android.view.MenuInflater
+import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
+import androidx.core.view.MenuProvider
 import androidx.navigation.NavController
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
@@ -9,7 +13,7 @@ import androidx.navigation.ui.navigateUp
 import androidx.navigation.ui.setupActionBarWithNavController
 import com.getwemap.example.map.databinding.ActivityInitialBinding
 
-class InitialActivity : AppCompatActivity() {
+class InitialActivity : AppCompatActivity(), MenuProvider {
 
     private lateinit var appBarConfiguration: AppBarConfiguration
     private lateinit var binding: ActivityInitialBinding
@@ -26,9 +30,26 @@ class InitialActivity : AppCompatActivity() {
         navController = findNavController(R.id.nav_host_fragment_content_initial)
         appBarConfiguration = AppBarConfiguration(navController.graph)
         setupActionBarWithNavController(navController, appBarConfiguration)
+
+        addMenuProvider(this)
     }
 
     override fun onSupportNavigateUp(): Boolean {
         return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
+    }
+
+    override fun onCreateMenu(menu: Menu, menuInflater: MenuInflater) {
+        menuInflater.inflate(R.menu.settings_menu, menu)
+    }
+
+    override fun onMenuItemSelected(menuItem: MenuItem): Boolean {
+        return when (menuItem.itemId) {
+            R.id.preferences -> {
+                findNavController(R.id.nav_host_fragment_content_initial)
+                    .navigate(R.id.action_Anywhere_to_SettingsFragment)
+                true
+            }
+            else -> super.onOptionsItemSelected(menuItem)
+        }
     }
 }