+++ map/src/main/java/com/getwemap/example/map/fragments/SettingsFragment.kt
@@ -1,0 +1,12 @@
+package com.getwemap.example.map.fragments
+
+import android.os.Bundle
+import androidx.preference.PreferenceFragmentCompat
+import com.getwemap.example.map.R
+
+class SettingsFragment : PreferenceFragmentCompat() {
+
+    override fun onCreatePreferences(savedInstanceState: Bundle?, rootKey: String?) {
+        setPreferencesFromResource(R.xml.root_preferences, rootKey)
+    }
+}